refactor(backend): extract service root path helpers
